- Moved title to annotation for Proxies. - Started looking a internationalisation support. Needs more thought.
Before this, the path I was going down was to set the title once during entity construction, based on an annotation or similar, however this won't work because the title may be different on each request for a entity if a different language is requested.

What makes more sense is to set the titles as above and then interpolate on serialisation. That would mean titles still only get set once, but we would translate on each request based on the requested language.