hmmm.... delete
Added action to delete proxies. It’s not pretty.

The crux of the problem is that from a java point-of-view, the deleting
action done on it’s container (i.e. for a Proxy, I should call delete
on Proxies with the name or the Proxy), but from a REST point-of-view
the delete should happen on the object.

I think this discrepancy might disappear when we start using
repositories. Will need to see.