More repository refactoring
The deserialisation is broken as the deserialised class refers to the
repository, which it does not have access to.

Interesting note: a client can effectively use the same repository
interface as the server, the difference being that the client’s
repository calls the rest services on the server. As I wrote.
Interesting.